[REFACTOR] #14 : 모듈 내 디렉토리명 변경 및 안쓰는 파일들 삭제
