+++ module-domain/src/main/kotlin/org/depromeet/team3/place/Menu.kt
@@ -1,4 +1,4 @@
-package org.depromeet.team3.menu
+package org.depromeet.team3.place
 
 import org.depromeet.team3.common.BaseTimeDomain
 import java.time.LocalDateTime
