+++ module-infra/src/main/kotlin/org/depromeet/team3/place/MenuJpaRepository.kt
@@ -1,4 +1,4 @@
-package org.depromeet.team3.menu
+package org.depromeet.team3.place
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
